+++ lab_2/src/core/List/ListSequence.hpp
@@ -7,7 +7,7 @@ class ListSequence : public Sequence<T> {
 private:
     void AppendInternal(T item);
     void PrependInternal(T item);
-    void InsertAtInternal(T item, int index);
+    void InsertAtInternal(T item, size_t index);
     void ConcatInternal(Sequence<T>* smth);
 
 protected:
@@ -16,9 +16,8 @@ protected:
     virtual ListSequence<T>* Clone() const = 0;
     virtual Sequence<T>* Instance() = 0;
 public:
-    //конструкторы
     ListSequence() : data_(new LinkedList<T>()) {}
-    explicit ListSequence(T* items, int count) 
+    explicit ListSequence(T* items, size_t count) 
         : data_(new LinkedList<T>(items, count)) {}
     explicit ListSequence(const LinkedList<T>& list) 
         : data_(new LinkedList<T>(list)) {}
@@ -28,21 +27,17 @@ public:
         delete data_;
     }
 
-    //Get-теры
     T GetFirst() const override;
     T GetLast() const override;
-    T Get(int index) const override;
-    int GetLength() const override;
-    Sequence<T>* GetSubsequence(int startIndex, int endIndex) const override;
+    T Get(size_t index) const override;
+    size_t GetLength() const override;
+    Sequence<T>* GetSubsequence(size_t startIndex, size_t endIndex) const override;
 
-    //set-теры
     Sequence<T>* Append(T item) override;
     Sequence<T>* Prepend(T item) override;
-    Sequence<T>* InsertAt(T item, int index) override;
+    Sequence<T>* InsertAt(T item, size_t index) override;
     Sequence<T>* Concat(Sequence<T>* smth) override;
 
-
-    //перегруз операторов
     ListSequence<T>& operator=(const ListSequence<T>& other) {
         if (this != &other) {
             delete data_;
@@ -53,5 +48,4 @@ public:
 };
 
 
-#include "ListSequenceGetMethods.tpp"
-#include "ListSequenceInsertMethods.tpp"
+#include "ListSequence.tpp"
